feat: load html for oss in ls (#570)
